requirements.txt
lyricsgenius 
ffmpeg-python 
pillow
python-dotenv